refactor: rename `__fixtures__` folder to `mocks`
this prevents clashing with jest
