+++ packages/notification-services-controller/notification-services/mocks/package.json
@@ -4,6 +4,6 @@
   "description": "",
   "license": "MIT",
   "sideEffects": false,
-  "main": "../../dist/NotificationServicesController/__fixtures__/index.cjs",
-  "types": "../../dist/NotificationServicesController/__fixtures__/index.d.cts"
+  "main": "../../dist/NotificationServicesController/mocks/index.cjs",
+  "types": "../../dist/NotificationServicesController/mocks/index.d.cts"
 }
