+++ packages/notification-services-controller/src/NotificationServicesController/processors/process-notifications.test.ts
@@ -2,9 +2,9 @@ import {
   processNotification,
   safeProcessNotification,
 } from './process-notifications';
-import { createMockFeatureAnnouncementRaw } from '../__fixtures__/mock-feature-announcements';
-import { createMockNotificationEthSent } from '../__fixtures__/mock-raw-notifications';
-import { createMockSnapNotification } from '../__fixtures__/mock-snap-notification';
+import { createMockFeatureAnnouncementRaw } from '../mocks/mock-feature-announcements';
+import { createMockNotificationEthSent } from '../mocks/mock-raw-notifications';
+import { createMockSnapNotification } from '../mocks/mock-snap-notification';
 import type { TRIGGER_TYPES } from '../constants/notification-schema';
 
 describe('process-notifications - processNotification()', () => {
