+++ packages/notification-services-controller/src/NotificationServicesController/processors/process-onchain-notifications.test.ts
@@ -15,7 +15,7 @@ import {
   createMockNotificationLidoWithdrawalRequested,
   createMockNotificationLidoWithdrawalCompleted,
   createMockNotificationLidoReadyToBeWithdrawn,
-} from '../__fixtures__/mock-raw-notifications';
+} from '../mocks/mock-raw-notifications';
 import type { OnChainRawNotification } from '../types/on-chain-notification/on-chain-notification';
 
 const rawNotifications = [
